Added functionality to ReportWindow.

+++ Vault/Core/Database/Data/Report.cs
@@ -13,27 +13,27 @@
         /// <summary>
         /// Gets or sets the number of passwords.
         /// </summary>
-        public long Total { get; set; } = 0;
+        public int Total { get; set; } = 0;
 
         /// <summary>
         /// Gets or sets the number of duplicated passwords.
         /// </summary>
-        public long Duplicated { get; set; } = 0;
+        public int Duplicated { get; set; } = 0;
 
         /// <summary>
         /// Gets or sets the number of weak passwords.
         /// </summary>
-        public long Weak { get; set; } = 0;
+        public int Weak { get; set; } = 0;
 
         /// <summary>
         /// Gets or sets the number of old passwords.
         /// </summary>
-        public long Old { get; set; } = 0;
+        public int Old { get; set; } = 0;
 
         /// <summary>
         /// Gets or sets the number of violated passwords.
         /// </summary>
-        public long Violated { get; set; } = 0;
+        public int Violated { get; set; } = 0;
 
         /// <summary>
         /// Gets or sets the timestamp.
@@ -43,13 +43,13 @@
         /// <summary>
         /// Initializes a new <see cref="Report"/> without id.
         /// </summary>
-        public Report(long total, long duplicated, long weak, long old, long violated, long timestamp)
+        public Report(int total, int duplicated, int weak, int old, int violated, long timestamp)
             : this(-1, total, duplicated, weak, old, violated, timestamp) { }
 
         /// <summary>
         /// Initializes a new <see cref="Report"/>.
         /// </summary>
-        public Report(int id, long total, long duplicated, long weak, long old, long violated, long timestamp)
+        public Report(int id, int total, int duplicated, int weak, int old, int violated, long timestamp)
         {
             Id = id;
             Total = total;
